Renamed TimeReference values to use PascalCase
C# naming conventions suggest to use PascalCase for Enums and kRPC
clientget assumes that kRPC services follow that convention. This change was
suggested in https://github.com/krpc/krpc/issues/490 to fix a problem with
incorrect Enum names being generated for C++ and Java.

This commit does not affect the users in any way since generated Enum names were
manually edited to use the correct format in the previous release.