Fix setting a target and calling a maneuver operation in the same frame
MechJeb has its own target controller which manages a reference to a current
target. This is updated once per frame in OnFixedUpdate() method. Using kRPC
API it is possible to bypass it - update the target and execute a maneuver
operation in the same frame. This resulted in the maneuver operation not knowing
about the target and raising an exception.

To fix it, Operation#MakeNodes() explicitly calls OnFixedUpdate() before
creating a maneuver node.